Move relativeTime logic to Timestmap container.
