+++ src/app/containers/Timestamp/timestampUtilities.test.js
@@ -1,8 +1,24 @@
-import formatUnixTimestamp from './timestampUtilities';
+import { formatUnixTimestamp, isValidDateTime } from './timestampUtilities';
 
 const timezone = 'Europe/London';
+const timestamp = 1539969006000; // 19 October 2018
 
 describe('Timestamp utility functions', () => {
+  describe('isValidDateTime', () => {
+    it('should return true if timestamp is valid', () => {
+      expect(isValidDateTime(timestamp)).toEqual(true);
+      expect(isValidDateTime(0)).toEqual(true);
+      expect(isValidDateTime(-30000000)).toEqual(true);
+    });
+
+    it('should return false if timestamp is invalid or missing', () => {
+      expect(isValidDateTime('foo')).toEqual(false);
+      expect(isValidDateTime(null)).toEqual(false);
+      expect(isValidDateTime(undefined)).toEqual(false);
+      expect(isValidDateTime()).toEqual(false);
+    });
+  });
+
   describe('formatUnixTimestamp', () => {
     it('should return BST for a BST timestamp', () => {
       // 31 May 2017 BST
